+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.117 ] - 2024/11/07
+
+  * Updated `SpotifyClient.GetArtistInfo` method to force all links in the `BioHtml` property to open in a new browser window / tab (e.g. added `target="_blank"`).
+
 ###### [ 1.0.116 ] - 2024/11/07
 
   * Updated `ArtistInfoTourEvent` model to include the concert link.